add how to feature (#11)
* add how to feature

* update workflow run trigger

* use reusable workflows

* dev reusable workflow

* add sample on how to dialog
