+++ .github/workflows/ci.yaml
@@ -1,5 +1,5 @@
 name: wotla CI
-on: pull_request
+on: workflow_call
 jobs:
   analyze:
     runs-on: ubuntu-latest
