+++ .github/workflows/firebase-deploy-dev.yml
@@ -3,16 +3,15 @@
 
 name: Deploy to Dev Hosting
 'on':
-  workflow_run:
-    workflows:
-      - 'wotla CI'
+  pull_request:
     branches:
       - 'main'
-    types:
-      - completed
 jobs:
+  test_and_analyze:
+    uses: ./.github/workflows/ci.yaml
   build_and_deploy:
     runs-on: ubuntu-latest
+    needs: test_and_analyze
     steps:
       - uses: actions/checkout@v2
       - uses: subosito/flutter-action@v2
