+++ .github/workflows/firebase-deploy-preview.yml
@@ -3,16 +3,16 @@
 
 name: Deploy to Preview Channel
 'on':
-  workflow_run:
-    workflows:
-      - 'wotla CI'
-    branches:
-      - '!main'
-    types:
-      - completed
+  pull_request:
+    branches-ignore:
+      - main
+
 jobs:
+  test_and_analyze:
+    uses: ./.github/workflows/ci.yaml
   build_and_preview:
     runs-on: ubuntu-latest
+    needs: test_and_analyze
     steps:
       - uses: actions/checkout@v2
       - uses: subosito/flutter-action@v2
