+++ lib/features/game/view/game_page.dart
@@ -6,9 +6,10 @@ import 'package:wotla/data/repositories/wotla_repository.dart';
 import 'package:wotla/features/game/bloc/game_bloc.dart';
 import 'package:wotla/features/game/bloc/game_event.dart';
 import 'package:wotla/features/game/bloc/game_state.dart';
+import 'package:wotla/features/game/widgets/answer_card.dart';
 import 'package:wotla/features/game/widgets/wotla_input.dart';
+import 'package:wotla/features/how_to/view/how_to_dialog.dart';
 import 'package:wotla/features/statistic/view/statistic_dialog.dart';
-import 'package:wotla/utils/const.dart';
 
 class GamePage extends StatelessWidget {
   const GamePage({Key? key}) : super(key: key);
@@ -46,33 +47,7 @@ class GameView extends StatelessWidget {
             showDialog(
               context: context,
               builder: (context) {
-                return AlertDialog(
-                  title: const Text('Cara Bermain'),
-                  content: Column(
-                    mainAxisSize: MainAxisSize.min,
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: const [
-                      Text(
-                        'Tebak WOTLA dalam ${WotlaConst.maxAttempt} kesempatan.',
-                      ),
-                      SizedBox(height: 8),
-                      Text(
-                        'Jawaban merupakan nama panggilan member JKT48 yang terdaftar pada web jkt48.com.',
-                      ),
-                      SizedBox(height: 8),
-                      Text(
-                        'Setelah jawaban dikirimkan, warna huruf akan berubah untuk menunjukkan seberapa dekat tebakanmu dengan jawabannya.',
-                      ),
-                      SizedBox(height: 8),
-                      Text(
-                        'Jika huruf berwarna hijau, maka huruf tersebut telah berada pada posisi yang tepat.',
-                      ),
-                      Text(
-                        'Jika huruf berwarna kuning, maka huruf tersebut terdapat pada jawaban, namun posisinya belum tepat.',
-                      ),
-                    ],
-                  ),
-                );
+                return const HowToDialog();
               },
             );
           },
@@ -131,34 +106,10 @@ class GameView extends StatelessWidget {
                               delegate: SliverChildBuilderDelegate(
                                 (context, index) {
                                   final history = state.history[index];
-                                  var charIndex = 0;
 
-                                  return Card(
-                                    color: const Color(0xFFFF99BB),
-                                    child: Padding(
-                                      padding: const EdgeInsets.all(8.0),
-                                      child: Row(
-                                        children: history.answerIdentifier
-                                            .toChars()
-                                            .map((char) {
-                                          final widget = Padding(
-                                            padding: const EdgeInsets.all(8.0),
-                                            child: Text(
-                                              history.answer[charIndex],
-                                              style: TextStyle(
-                                                color: char == 'X'
-                                                    ? Colors.black
-                                                    : char == '+'
-                                                        ? Colors.yellowAccent
-                                                        : Colors.green,
-                                              ),
-                                            ),
-                                          );
-                                          charIndex++;
-                                          return widget;
-                                        }).toList(),
-                                      ),
-                                    ),
+                                  return AnswerCard(
+                                    answer: history.answer,
+                                    answerIdentifier: history.answerIdentifier,
                                   );
                                 },
                                 childCount: state.history.length,
