+++ lib/utils/string_extension.dart
@@ -1,0 +1,11 @@
+extension StringExtension on String {
+  List<String> toChars() {
+    final chars = <String>[];
+
+    for (int i = 0; i < length; i++) {
+      chars.add(this[i]);
+    }
+
+    return chars;
+  }
+}
